requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
progressbar2==4.4.2
